requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
django-filter==21.1
django-cors-headers==3.10.0
sendgrid==6.9.6
gunicorn==20.1.0
requests
facebook-sdk==3.1.0
easy-thumbnails==2.8.1
pyjwt==2.2.0